apps/vibe-justice/backend/requirements.txt
# Python Version Requirement: 3.11-3.13 (NOT 3.14+)
# Reason: chromadb dependency onnxruntime not yet compatible with Python 3.14
# Recommended: Python 3.12.x for optimal compatibility

# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.1
pydantic==2.9.2
pydantic-settings==2.6.1

# AI/ML
openai==1.58.1
chromadb==1.3.7
pypdf==5.1.0
python-docx==1.1.2

# Database
sqlmodel==0.0.22
sqlalchemy==2.0.36
alembic==1.14.0

# Utilities
python-dotenv==1.0.1
python-multipart==0.0.17
psutil==6.1.1
httpx==0.28.1
requests==2.32.3

# Web Search (no API key required)
duckduckgo_search==7.5.3

# Rate Limiting & Caching
slowapi==0.1.9
redis==5.2.1
hiredis==3.0.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0

# OCR & Document Processing
pytesseract==0.3.13
pdf2image==1.17.0
Pillow==11.0.0

# File System Monitoring
watchdog==6.0.0

# Task Queue (for batch processing)
celery==5.4.0

# Development
ruff==0.8.4
mypy==1.13.0

# Packaging
pyinstaller==6.12.0
